refactor: move color control
