+++ packages/controls/src/index.ts
@@ -1,6 +1,7 @@
 export * from './common'
 
 export * from './checkbox'
+export * from './color'
 export * from './number'
 export * from './text-input'
 export * from './text-area'
