+++ packages/runtime/src/controls/color.ts
@@ -1,39 +1,10 @@
 import { match } from 'ts-pattern'
-import { ControlDataTypeKey, ColorData } from '@makeswift/controls'
+import { ColorControlData, ColorControlDataTypeKey, ColorControlDataTypeValueV1 } from '@makeswift/controls'
 
 import type { CopyContext } from '../state/react-page'
 
-export const ColorControlDataTypeKey = ControlDataTypeKey
 
-export const ColorControlDataTypeValueV1 = 'color::v1'
-
-export type ColorControlDataV0 = ColorData
-
-export type ColorControlDataV1 = ColorControlDataV0 & {
-  [ColorControlDataTypeKey]: typeof ColorControlDataTypeValueV1
-}
-
-export type ColorControlData = ColorControlDataV0 | ColorControlDataV1
-
-export const ColorControlType = 'makeswift::controls::color'
-
-type ColorControlConfig = {
-  label?: string
-  labelOrientation?: 'horizontal' | 'vertical'
-  defaultValue?: string
-  hideAlpha?: boolean
-}
-
-export type ColorControlDefinition<C extends ColorControlConfig = ColorControlConfig> = {
-  type: typeof ColorControlType
-  config: C
-  version?: 1
-}
-
-export function Color<C extends ColorControlConfig>(config = {} as C): ColorControlDefinition<C> {
-  return { type: ColorControlType, config, version: 1 }
-}
-
+// @todo: move this into controls package
 export function copyColorData(
   value: ColorControlData | undefined,
   context: CopyContext,
